Adjust plugin regex logic
Previously we used a simple string 'in' to match the command. This was
unreliable as it could match other parts of output e.g. file names etc.
By making the command a regex it can be more specific. We use search
rather than match so it's more flexible and if you have no regex
knowledge you can configure it like a normal string match.

Prompt matching also uses search now so you don't need to write a regex
for the whole line and could use normal strings if you prefer.

We also handle if command isn't specified properly, before we'd just
give the plugin one line rather than the whole output. Now we give the
whole output

Regex's are compiled outside handle to make it slightly more efficient,
compiling every time you hit the shortcut is a bit of a waste